overleaf-mcp/requirements.txt
fastmcp>=0.2.0
uvicorn>=0.20.0
starlette>=0.27.0
